Allow AWS configuration to come from other places than just environment variables
The AWS client library doesn't have to use environment variables to get credentials and configuration.

See https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/setting-credentials-node.html for some of the ways that credentials can be provided.

So, we should not abort if the environment variables are missing, we can let the AWS SDK throw an error if there really are no credentials to be found.